ecommerce/ecommerce/requirements.txt
channels-redis
django-allauth
django-apscheduler
django-bootstrap-v5
django-bootstrap5
django-bootstrap5-form
django-celery-beat
django-celery-results
django-ckeditor
django-constance
django-material
django-picklefield
django-redis
django-rest-auth
django-viewflow
djangorestframework-stubs
drf-nested-routers
flake8
install
pep517
pillow
pip-chill
pyhamcrest
pyopenssl
pytest-django
pyyaml
service-identity
sorl-thumbnail
sphinx-rtd-theme
twilio
